+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [#911](https://github.com/hyperf/hyperf/pull/911) Added support for crontab task run on one server.
 - [#913](https://github.com/hyperf/hyperf/pull/913) Added `Hyperf\ExceptionHandler\Listener\ErrorExceptionHandler`.
 - [#931](https://github.com/hyperf/hyperf/pull/931) Added `strict_mode` for config-apollo.
+- [#933](https://github.com/hyperf/hyperf/pull/933) Added plates template engine for view.
 
 ## Fixed